requirements.txt
Django==2.2.3
djangorestframework==3.9.1
pytz==2019.2
sqlparse==0.3.0
requests==2.22.0